Added new startup procedure, configured logging and added (using Srubin's code as a base) addition/removal of members.
Startup now only requires the number of nodes to be launched. One of the nodes will be designated as the founder of the group, and automatically assumes leadership. Other nodes begin in the 'Outsider' state. They will listen for heartbeats and ask to join the group.

The leader will add/remove nodes one at a time, following the Raft protocol. Logs are maintained for each member of the group, in separate files.

NB) This code contains a hard-coded failure case: Member 3 will always fail to send their fourth ACK, and should be removed by the leader.

Code is still heavily untested - probably has a lot of issues that haven't been accounted for.